Add files via upload
Performed web scrapping to get a list of available appartments in Hartford, CT when I had to relocate to Hartford for my internship. The major thing that bothers everyone is to find a suitable accomodation at a reasonable price. Being a student who would wish to live at a place in a good neighborhood and at a reasonable price, I developed a web scraper in python to get a list of all the appartments in the vicinity in a given price range, listing the necessarry details - appartment name, url, contact details, price, availability status.